Added Sphinx orphan directive to suppress toctree errors
If you use the include directive in a Sphinx projects, the build
generates an error WARNING: document isn’t included in any toctree.
This happens because the file is read twice, once where the content is
included, and again in the stand-alone file where the content is
stored.  To suppress this error, you can add the  :orphan: directive on
the first line.